backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-cors-headers
django-environ
dj-rest-auth
django-allauth
drf-yasg
channels
channels_redis
celery
django-celery-beat
PyJWT>=2.0.0
cryptography>=41.0.0
djangorestframework-simplejwt
requests
django-extensions
requests-oauthlib
dj-rest-auth[with_social]



